Split arg parsing logic from the app entry point
Problem
-------

The CLI argument parsing logic needs to be reasoned about
separately from the main entry point of the application.

Solution
--------

Split an args.hpp and args.cpp out from run.cpp.
